requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.7.0
google-generativeai>=0.8.0
typer>=0.12.0
rich>=13.7.0
